+++ LookupAnything/Framework/Fields/ItemGiftTastesField.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Linq;
+using Microsoft.Xna.Framework;
 using Pathoschild.Stardew.LookupAnything.Framework.Constants;
+using Pathoschild.Stardew.LookupAnything.Framework.Models;
 
 namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
 {
@@ -13,10 +15,12 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
         /// <summary>Construct an instance.</summary>
         /// <param name="gameHelper">Provides utility methods for interacting with the game code.</param>
         /// <param name="label">A short field label.</param>
+        /// <param name="itemID">The item ID for which to show gift tastes.</param>
         /// <param name="giftTastes">NPCs by how much they like receiving this item.</param>
         /// <param name="showTaste">The gift taste to show.</param>
-        public ItemGiftTastesField(GameHelper gameHelper, string label, IDictionary<GiftTaste, string[]> giftTastes, GiftTaste showTaste)
-            : base(gameHelper, label, ItemGiftTastesField.GetText(giftTastes, showTaste)) { }
+        /// <param name="onlyRevealed">Only show gift tastes the player has discovered for themselves.</param>
+        public ItemGiftTastesField(GameHelper gameHelper, string label, int itemID, IDictionary<GiftTaste, GiftTasteModel[]> giftTastes, GiftTaste showTaste, bool onlyRevealed)
+            : base(gameHelper, label, ItemGiftTastesField.GetText(giftTastes, showTaste, onlyRevealed)) { }
 
 
         /*********
@@ -25,13 +29,34 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Fields
         /// <summary>Get the text to display.</summary>
         /// <param name="giftTastes">NPCs by how much they like receiving this item.</param>
         /// <param name="showTaste">The gift taste to show.</param>
-        private static string GetText(IDictionary<GiftTaste, string[]> giftTastes, GiftTaste showTaste)
+        /// <param name="onlyRevealed">Only show gift tastes the player has discovered for themselves.</param>
+        private static IEnumerable<IFormattedText> GetText(IDictionary<GiftTaste, GiftTasteModel[]> giftTastes, GiftTaste showTaste, bool onlyRevealed)
         {
             if (!giftTastes.ContainsKey(showTaste))
-                return null;
+                yield break;
 
-            string[] names = giftTastes[showTaste].OrderBy(p => p).ToArray();
-            return string.Join(", ", names);
+            // get data
+            string[] names =
+                (
+                    from entry in giftTastes[showTaste]
+                    orderby entry.Villager.Name ascending
+                    where !onlyRevealed || entry.IsRevealed
+                    select entry.Villager.Name
+                )
+                .ToArray();
+            int unrevealed = onlyRevealed
+                ? giftTastes[showTaste].Count(p => !p.IsRevealed)
+                : 0;
+
+            // build result
+            if (names.Any())
+            {
+                yield return new FormattedText(string.Join(", ", names));
+                if (unrevealed > 0)
+                    yield return new FormattedText(L10n.Item.UndiscoveredVillagersAppend(count: unrevealed), Color.Gray);
+            }
+            else
+                yield return new FormattedText(L10n.Item.UndiscoveredVillagers(count: unrevealed), Color.Gray);
         }
     }
 }